+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * [EACQAPW-5262] CardsController Tests
 * [EACQAPW-5218] AddCardController final tests
 * [EACQAPW-5281] CardRequisitesMasksResolver Tests
+* [EACQAPW-5292] CardRequisitesValidator Tests
 
 ### Fixed